// web/PirateCardsLayaGame/client/Game/src/GameModule/MenuIndexId.ts
export enum MenuIndexId 
{
    //-------招募---------------
    /** 普通招募 */
    GashaponGeneral  = 200000,

    /** 高级招募 */
    GashaponAdvanced = 200001,

    //-------英雄---------------
    /** 英雄基础 */
    HeroBase  = 201000,

    /** 英雄技能 */
    HeroSkill = 201001,

    /** 英雄装备 */
    HeroEquip = 201002,

    /** 英雄技能 */
    HeroStar  = 201003,

    //-------背包---------------
    /** 背包普通 */
    BagNormal = 202000,

    /** 背包碎片 */
    BagPieces = 202001,

    /** 背包装备 */
    BagEquip  = 202002,

    //-------装备---------------
    /** 装备强化 */
    EquipStrenghen = 203000,

    /** 装备升星 */
    EquipStar      = 203001,

    //-------战魂---------------
    /** 战魂觉醒 */
    DraugAwaken = 204000,

    /** 战魂装配 */
    DraugWeared = 204001,

    /** 战魂分解 */
    DraugDecompose = 204002,

    //-------征战---------------
    /** 剧情关卡 */
    SectionStory = 207000,

    /** 精英关卡 */
    SectionElite = 207001,


    //-------任务---------------
    /** 日常任务 */
    QuestDaily  = 300002,

    /** 成长任务 */
    QuestGrowUp = 300001,



    //-------角色---------------
    /** 角色信息 */
    PlayerInfo = 503000,

    /** 角色战队 */
    PlayerTeam = 503001


}